requirements.txt

fastapi
uvicorn
python-docx
